Remove SplitPane from `common/index` (#6021)
[skip ci]
