+++ app/javascript/components/common/index.ts
@@ -10,7 +10,6 @@ export { AvatarGroup, AvatarGroupProps } from './AvatarGroup'
 export { ProminentLink } from './ProminentLink'
 export { ProcessingStatusSummary } from './ProcessingStatusSummary'
 export { default as Expander } from './Expander'
-export { SplitPane } from './SplitPane'
 export { Tab, type TabContext } from './Tab'
 export { SingleSelect } from './SingleSelect'
 export { MultipleSelect } from './MultipleSelect'
